wip: Add 'panic mode' to libtorrent.patch
This will connect any peers taking more than 5 seconds to download an
'important' piece which the client needs very soon for playback.

This works very well but is maybe a little too harsh. Ideally we would just
put additional peers on downloading the same piece. Then we could just add
a peer every 5 seconds or so.